+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.5.0 - TBD
 - Added a parser for Virtual Judge
 - Fixed bugs in the ACMP, ECNU Online Judge, UVa Online Judge, ICPC Live Archive and AtCoder parsers (thanks [@m1kit](https://github.com/m1kit) for the AtCoder fix)
-- Add "interactive" to the output format which specifies whether the problem is interactive or not
+- Added "interactive" to the output format which specifies whether the problem is interactive or not
+- Removed the regex input option
 
 # 2.4.3 - December 14th, 2018
 - Fixed bugs in the ACMP and the Toph parsers